[test:36] making changes to remove duplicates
